pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mp-weixin-skills"
version = "0.1.0"
description = "微信公众号文章发布工具"
readme = "SKILL.md"
requires-python = ">=3.10"
authors = [
    {name = "Claude AI"}
]
dependencies = [
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "python-docx>=1.0.0",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
mp-weixin = "scripts.cli:main"

[tool.setuptools.packages.find]
where = ["scripts"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
